FixedBugs

+++ app/src/main/java/com/krpvartstudio/sshine/business/model/MainHourListModel.kt
@@ -9,7 +9,7 @@ data class MainHourListModel(
     val pop: Double,
     val pressure: Int,
     val temp: Double,
-    val uvi: Int,
+    val uvi: Double,
     val visibility: Int,
     val weather: List<Weather>,
     val wind_deg: Int,
